Added JavaDoc comments to all of Shelly and My classes. Also renamed ScannableCodeUnitTest.java to ScannedCodeUnitTest.java.
