+++ app/src/main/java/com/example/qrchive/Classes/FirebaseWrapper.java
@@ -24,6 +24,13 @@ import java.util.concurrent.ExecutionException;
 
 // In the ScannedCodes collection on firebase, the userDID, location, hasLocation, hash can ...
 // ... uniquely identify a QR code
+
+/**
+ * A wrapper class for Firebase that allows developers to easily manipulate the Firebase DB.
+ *
+ * @author Shelly
+ * @version 1.0
+ */
 public class FirebaseWrapper {
     public FirebaseFirestore db;
     private String myDeviceID;
@@ -32,12 +39,23 @@ public class FirebaseWrapper {
     private HashMap<String, ArrayList<ScannedCode>> scannedCodesDict = new HashMap<>();
     private ArrayList<String> users; //part 4
 
+    /**
+     * FirebaseWrapper constructor, instantiates a single instance of the FirebaseWrapper class.
+     *
+     * @param myUserDID is the device ID of the user we want to operate on.
+     */
     public FirebaseWrapper(String myUserDID) {
         this.db = FirebaseFirestore.getInstance();
         this.myUserDID = myUserDID;
         this.refreshScannedCodesForUser(myUserDID);
     }
 
+    /**
+     * refreshScannedCodesForUser will re-initialize the scannedCodesDict with all of the users
+     * scanned codes.
+     *
+     * @param userDID is the user device ID we are trying to fetch the codes for.
+     */
     public void refreshScannedCodesForUser(String userDID) {
         ArrayList<ScannedCode> scannedCodes = new ArrayList<>();
         db.collection("ScannedCodes").whereEqualTo("userDID", userDID).get()
@@ -68,15 +86,29 @@ public class FirebaseWrapper {
     }
 
 
+    /**
+     * getMyUserDID is a getter function for myUserDID.
+     *
+     * @return Returns private attribute myUserDID.
+     */
     public String getMyUserDID() {
         return myUserDID;
     }
 
+    /**
+     * getScannedCodesDict is a getter function for scannedCodesDict.
+     *
+     * @return Returns private attribute scannedCodesDict.
+     */
     public HashMap<String, ArrayList<ScannedCode>> getScannedCodesDict() {
         return scannedCodesDict;
     }
 
-    // Deletes the code from firebase and also updates the hashmap
+    /**
+     * deleteCode will delete a ScannedCode from the firebase DB and maintain the scannedCodesDict.
+     *
+     * @param scannedCode is the code we wish to delete from the DB.
+     */
     public void deleteCode(ScannedCode scannedCode) {
 //        Tasks.await()
         db.collection("ScannedCodes").document(scannedCode.getScannedCodeDID())
@@ -89,6 +121,10 @@ public class FirebaseWrapper {
 
     }
 
+    /**
+     * getUsers is a getter function for users.
+     * @return Returns private attribute users.
+     */
     public ArrayList<String> getUsers() {
         return users;
     }
